requirement.txt
comtypes==1.1.9
pycss==0.0.1
pyperclip==1.8.2
pypiwin32==223
pyttsx3==2.90
pywin32==300
